refractor: CustomIcon into It's own component
